requirements.txt
altair==5.3.0
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
catboost==1.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
entrypoints==0.4
extra-streamlit-components==0.1.71
Faker==25.8.0
favicon==0.7.0
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
htbuilder==0.6.2
idna==3.7
imbalanced-learn==0.12.2
Jinja2==3.1.4
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==10.2.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
prometheus_client==0.20.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pymdown-extensions==10.8.1
pyodbc==5.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.13.1
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.35.0
streamlit-authenticator==0.3.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-dynamic-filters==0.1.6
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.3
watchdog==4.0.1
wheel==0.43.0